addrxlat/sys: Add hardware mapping
This mapping corresponds to how hardware (CPU) translates virtual
addresses. Keep in mind that this is not a simple wrapper around
ADDRXLAT_SYS_METH_PGT, because:

 1. Some areas may be undefined (e.g. the non-canonical region on
    x86_64), which is reflected in the map as NULL method.

 2. Some CPUs do not implement page tables in hardware (e.g. IBM
    POWER).

NB: There is currently no code to do interpret HTAB on Power ISA,
but if it is added later, it will be referenced by this mapping.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>