addrxlat_map_set: Correctly handle empty but allocated maps
The function only checks if the map is unallocated, but it is
possible to have an allocated map with n=0, e.g. after calling
addrxlat_map_clear() on it.

Note that the reverse condition (!map || !map->n) should be used
later after calling realloc(), but since first is always set to
a non-NULL value, except if the first conditional is false, it can
be optimized to if (!first).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>