addrxlat/x86_64: Translate machine addresses when probing layout
This did not matter for the non-Xen case, because machine addresses
are equal to kernel physical addresses, but after adding support for
Xen, the checks for direct mapping and kernel text mapping must use
kernel physical addresses.

Funnily, the directmap bug cannot triggered on 2.6.31, because
is_directmap() success is not checked for 2.6.31+ (the existence of
the VMEMMAP array is sufficient). That's why I added the new case for
2.6.27 under Xen (with no plans to check 2.6.31+).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>