addrxlat: Improve addrxlat_ctx_err()
First of all, let's make the error string grow dynamically. This
is good, because it can grow rather large if something fails, but
it is usually not needed at all.

There is a small catch: allocation failures. If an allocation fails
for the caller, it is likely that it will also fail during error
handling, in turn giving no clue to the library user. So, let's
keep a smaller fallback buffer just for these cases (or when the
error message is short enough to fit).

Now, if the message must be stored in the fallback buffer, it may
have to be truncated. This was indicated with an ellipsis, but an
ellipsis takes up three characters of the (precious) space. Use
a '<' character to indicate that the message continues to the left
and a '>' character to indicate that it continues to the right.

Why would a message be ever truncated on the right? This may happen
if allocation fails. The function must then use a temporary buffer
on stack (called lbuf in the code) to get the full message, because
vsnprintf() can only truncate the end of the message. This buffer is
currently sized to ERRBUF, but it does not have to be the same size
as the fallback buffer.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>