addrxlat: Re-design handling of internal-only symbols
Some symbols are used only internally by the library, but from
different object files, so they must be extern. To protect the
global namespace in a static link, these symbols are prefixed
with "_libaddrxlat_priv". However, I don't want to use such
cumbersome names in the source code, so they are remapped
automatically.

So far the mapping has been done by the preprocessor. This is
suboptimal, because the preprocessor changes all occurences of
that symbol (e.g. arguments or structure field names with the
same name).

Instead, let's exploit GCC's ability to assign an arbitrary
name to a symbol with the __asm__() directive in the declaration.

As a side note, the prefix is now shortened to "_addrxlat_priv".
This should still be good enough to avoid potential clashes with
other libraries.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>