Re-design addrxlat walk/step API (take two)
The biggest challenge is that the API should expose enough details
to implement a step-by-step "vtop" command while providing a simple
interface for the casual user, all that without exploding the number
of exported functions.

These are the design decisions:

1. All translations can be performed in steps. Sometimes there is
   only one step.
2. The step state is transparent to library users and part of the
   library API and ABI. This allows efficient access to translation
   details (no need for accessors).
3. The first step is different from the subsequent steps, because
   it does not use any previous content of the step state. Instead,
   the first step causes an initialization of the step state.
   This step is performed by addrxlat_launch_* functions.
4. The step state can be modified after launching the translation.
   This allows to initialize all values on launch without having
   to pass all potentially useful values as arguments to the launch
   function.

NOTE: An important change was made in the internal API. The first
step function of a translation method should never return
addrxlat_continue. It is now implied that the first step is
followed by at least one call to the next step function.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>