addrxlat: Pass also translation system to read32/read64
Main motivation is that reading an address may itself require
address translation to something that the read method can handle.

The downside is that a few initialization routines have to provide
a dummy translation step state. A clean solution would be a new
structure which contains just the context and system and is
embedded in addrxlat_step_t, but it's overkill for the two places
that actually need it...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>