tests: Add p2m translation for Linux root page table
The missing p2m translation has been masked by the fact that
get_physaddr() doesn't check whether the result of KV->PHYS is
a machine physical address or a kernel physical address.

So, I've set up the table to map PFN 0x1e10 to MFN 0x1e10. This
is certainly not optimal, as it may hide other missing translations,
but let's leave it like that for now.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>