addrxlat: Implement read callback capabilities
The trouble with read callbacks is that they may not be able to
process any address space directly, calling back into libaddrxlat
to translate, which in turn may need a read callback, etc...

The only clean solution is to let the library know which address
spaces can be read without further translation.

NB: There is currently no code to check for infinite recursion,
so this commit alone does not solve the original problem.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>