Tell libaddrxlat to perform Xen address translation if needed
If a Xen dump is detected, kernel physical addresses do not match
machine physical addresses. There is support for this in libaddrxlat,
but the library itself cannot detect whether such translation is
necessary.

Note that auto-translated domains make use of hardware capabilities,
and this translation is transparent to libaddrxlat. However, dumps
of the whole system will always need it, because machine addresses
then correspond to bare metal, regardless of what hw virtualization
was active for Dom0.

This commit enables correct handling of Dom0 dumps.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>