python: Use PYTHON build flags correctly
First, the module is written in C, not C++, so *_CFLAGS is used.
Second, use PYTHON_LIBS for link. GNU ld does not really complain
when building a shared object, and Python happily loads the object
even if libpython is not listed as NEEDED (because it is pulled in
by the executable itself), but it still feels wrong to rely on such
assumptions.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>