x86_64: Improve Xen vtop initialization
Since the root page table address is (once again!) provided
as a virtual address, there is a chicken-and-egg problem, so
initialization must be done a lot more carefully to make sure
we can use all available information.

Without this patch, Xen vtop initialization usually fails,
because the root page table is never usable.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>