Fix various issues related to distributed include files
First and foremost, the public header files are generated by
./configure, so that it can include version information. It is
not possible to achieve this with pre-processor macros, because
they are only defined during package build (i.e. unavailable to
library users after it is installed).

Consequently, these header files should not be distributed, as
they are generated objects, so add a "nodist_" prefix to Automake.

Finally, since these files are not distributed, they are not found
under top_srcdir but under top_builddir. This makes a difference
for VPATH builds, which would be failing otherwise.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>