Improve addrxlat_launch API
Try to clean up some of the mess around making the first translation
step. In particular, provide a low-level interface for making the
first step (addrxlat_launch).

Second, it turns out that addrxlat_launch_meth would add so little
on top of a raw addrxlat_launch() call that it is not worth keeping
it.

Third, modify addrxlat_launch_map() to match the new paradigm:
Sset all fields in addrxlat_step_t prior to making a launch call;
do not adjust it afterwards.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>