Introduce addrxlat_op
Use this function to run an operation on any address making any needed
translations to convert it to one of the address spaces supported by
the operation.

The main use should be for read accessors, which may not be able to
operate on any address space, but may still support more than one
address space. Using addrxlat_by_sys() is difficult in these cases,
because the caller would have to decide which target address space
should be used (or even retry mulitple times if needed).

The addrxlat_op() function takes a bitmask of all address spaces that
can be handled by the operation and performs the best translation
automatically, so the logic does not have to be re-implemented in
every application.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>