Use page address instead of PFN in struct page_io
This may slightly reduce the overhead of address translation, but the
main point is preparation for storing a addrxlat_fulladdr_t instead.

Note that shifting right and then left is not a no-op; it clears the
lower bits, so replace right shifts by a call to page_align(). It is
possible to skip this step after address translation, because the
translation operates on pages (so a page-aligned address is always
translated to a page-aligned address).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>