Convert libkdumpfile address conversions to addrxlat_op
Use a full address for page I/O and use addrxlat_op() to convert
the address space to something that is understood by the format's
read_page() method.

Note that only after this commit, kdump_read() starts taking the
address space seriously, so for example no address translation is
possible until the OS type is set (e.g. to "linux"). This broke the
dumpdata utility from the test suite.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>