s390x: initialize vtop translation before reading low core
Low core is in kernel physical memory, but this is identical to
machine physical on s390x (well, until there is some support for
complete z/VM dumps).

Let's not change the address space to ADDRXLAT_MACHPHYSADDR, because
that wouldn't be correct. Instead, set up a temporary mapping before
attempting to read lowcore info.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>