python: Discard exceptions from get_symbol_func
If a symbol cannot be found, the callback function is supposed to
raise an exception. The glue code turns the exception into an error
code which may be ignored by libkdumpfile. However, the outstanding
exception is not properly handled by Python and it may cause trouble
at a random point later.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>