Remove kdump_vtop_init() from the API
Address translation is now always initialized. The original concern
is addressed by depending on a valid OS type. Setting up address
translation for an unknown OS should be fast enough.

Effectively, calls to kdump_vtop_init() would be replaced with setting
the ostype attribute to "linux". Note that it is also possible to set
it to other values (currently only "xen").

Signed-off-by: Petr Tesarik <ptesarik@suse.com>