Try to initialize version code before vtop_init()
The version code may be available early (e.g. if it is stored in the
dump file header, or parsed from VMCOREINFO). If so, then it might be
useful to libaddrxlat initialization to enable fallback mechanisms of
determining virtual memory layout.

Note that the second call to linux_version_code() and xen_version_code()
should be preserved for the case that linux.uts.release is set later by
update_linux_utsname().

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: Petr Tesarik <ptesarik@suse.com>