addrxlat_launch: Return ADDRXLAT_ERR_NOMETH if method is NULL
Return an error code instead of segfaulting. This should make using
the library a bit less frustrating, because meth must be set to a
valid value for some calls (addrxlat_launch) but not for other calls
(addrxlat_launch_map).

Note that context is not checked, because context must be set to a
valid value for all calls involving addrxlat_step_ctx_t.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>