python/addrxlat: Relax argument checks in __init__()
PyArg_ParseTupleAndKeywords() performs a strict check on the
arguments, e.g. any extra arguments are treated as an error. This
makes it very difficult to write derived classes, because one would
have to override __new__() as well as __init__() just to silence the
TypeError exceptions. That's overkill...

Instead, drop the ability to accept keyword arguments in the low-level
module, and do not complain about extra positional arguments.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>