python/addrxlat: review actual use of description parameter locations
It turns out that at most two locations are needed. This structure is
not part of any ABI, so this number can be increased if necessary.

Another observation: There is always at least one location, and the
first param location always covers the whole raw parameter. This can
be leveraged to simplify desc_param_len().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>