Add ADDRXLAT_CUSTOM
Allow library users to specify custom translation step functions.
This translation kind can be used if a library user has an
alternative method for a specific translation but wants to use the
rest of the libaddrxlat infrastructure. For example, Xen machine
addresses and kernel physical addresses can be translated using a
special ELF section for Xen DomU dumps.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>