python/addrxlat: Provide sequence interface to Map()
A translation map contains a sequence of ranges, so it can be treated
as a sequence, instead of providing len() and get_range() methods.

It breaks the rule to keep the _addrxlat API as low-level as possible,
but in the end, this change leads to less code, wile being more
useful at the same time.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>