python/addrxlat: Use _C_POINTER and co-operative __init__()
Arguably, this is the cleanest way to enable subclassing. All classes
now use super(), chopping off whatever arguments they accept and
passing the rest to the next method.

This was necessary to add a special argument called _C_POINTER to the
_addrxlat base classes. The object carries a pointer to an existing
C object. A new object is not created in this case.

Initialization from a C type is done by *_FromPointer. The trouble
with these functions is that they do not call __new__() or __init__(),
instead allocating directly using tp_alloc() and setting some
structure fields. This is very confusing if a derived class relies on
additional initialization by __init__()... For example, note that
*_FromPointer no longer touch the convert slot, but it is still
initialized properly by subclasses in addrxlat.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>