python/addrxlat: Remove Operator.result
The current implementation holds a reference to the value returned
by callback(), which may be a large object. Instead, let's return a
tuple with the status code and the actual result.

Note that addrxlat.Operator takes care of exception handling, so it
can return the value unpacked (the status code is not useful then).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>