addrxlat: Add a hook for modifying callback settings
Such a hook is necessary to allow tracking changes to the callback
settings (e.g. through a Python object).

The hook is called whenever the callbacks are set, giving the hook
a chance to modify the result. Note that the original values are
also kept and returned by addrxlat_get_cb(), so these modifications
are hidden from normal API users. However, it also means that a new
function is needed to install the hook.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>