python/addrxlat: Store Context callbacks to slots
The problem with function descriptors is that they are read-only.
Sure, only the base class is affected, because derived types can
define their own __dict__, which can be used to override a method
descriptor. This is in fact the case for addrxlat.Context() already.

However, it is not possible to initialize the callback functions
correctly in the base type if it is created from a C pointer.

While it would be possible to add a __dict__ to the base type, I
chose to add three new slots, because it's more efficient than
searching the dictionary (and the callback might be used quite often
for page-table translation).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>