python/addrxlat: Improve exception handling for chained Contexts
To improve error reporting, handle the stored Python exception in
the next object if possible.

Note that this is not perfect, because the information is lost if
there is another hook between the Context hook that raised the
exception and the calling Context, but this patch can only make the
situation better, never worse.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>