python/addrxlat: Use co-operative super() in exceptions
This "advanced Python magic" to get the exception type before it
is created works, because the __init__() function is defined in a
closure and "cls" gets stored as a reference, so when it is set
later in the same function, it refers to the newly created type.

BTW the same mechanism is employed for the value of status already.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>