addrxlat: Rename ADDRXLAT_MAXLEVELS to ADDRXLAT_FIELDS_MAX
The current name is confusing, because it appears to be off by one.
For example, the current value (8) is enough only for 7-level paging,
because the first array element is used for page offset.

Now that "the thing" is finally called address bit fields, this
constant should be renamed, too.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>