addrxlat: Rename addrxlat_sys_init to addrxlat_sys_os_init
The idea came through Python bindings, where init() was very similar
to __init__(), but could not be really merged, as it is possible to
manually fill in the mapping in an addrxlat_sys_t object without
calling sys_init() on it.

The main point is that the automatic initiliazation uses some
hard-wired heuristics for Linux and Xen. So, let's just add "os_"
into the name to make it obvious that this method is somehow related
to OS details.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>