python/addrxlat: Helper to strip off arguments in cooperative classes
Unfortunately, the default Python C API does not provide a function
to strip off part of the arguments and pass the rest down to the next
method.

All PyArg_Parse* methods assume that the caller consumes _all_ its
arguments, but this makes it very difficult to add new arguments in
derived classes, because each class would have to override both
__init__() and __new__() to make sure that __new__() does not get
those extra arguments...

Let's use this helper instead of nth_arg(), which was the previous
attempt to achieve some flexibility, but didn't cope with keyword
arguments.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>