python/addrxlat: Allow overriding Context callbacks in a subclass
Storing the callbacks in a data slot does not give enough flexibility
to API users. The main downside is that new methods are always stored
in __dict__, but the dictionary is not checked for data descriptors.
Consequently, it is not possible to define a callback simply by
defining a cb_*() method in a derived class.

In other words, this will not work:

class mycontext(Context):
    def cb_read64(self, addr):
        # My cool 64-bit reader here

It is possible to work around it, e.g. by writing:

class mycontext(Context):
    def cool_cb_read64(self, addr):
        # My cool 64-bit reader here

    def __init__(self, *args, **kwargs):
        super(mycontext, self).__init__(*args, **kwargs)
        self.cb_read64 = self.cool_cb_read64

However, this is quite confusing, so for a type that is supposed to
serve as a base class, let's store the value in __dict__ if necessary.

Note that a _METHOD_ descriptor for the default behaviour can be
provided: if there is a non-data descriptor, Python will also search
__dict__ and prefer the value from __dict__ if found.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>