Rename addrxlat_map_dup() to addrxlat_map_copy()
The original motivation was to align with Python naming (compare this
to e.g. dict.copy), but dup[licate] may actually give the wrong idea
of the operation (compare this to the dup() system call). The result
of addrxlat_map_copy() is an independent shallow copy of the original
map, much more like Python's dict.copy() than a POSIX dup(2).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>