python/addrxlat: Fix ctx hook deinstallation
This is what I call a braino.

It's not possible to uninstall the hook from the middle of the chain
using addrxlat_ctx_install_cb_hook(), because the new hook is called
with the current (effective) callbacks, which is then pushed down the
stack and stays persistent. :-(

Instead, the caller should re-install the existing original callbacks,
making sure that they do not get modified by the currently installed
hook (code already works that way after setting self->ctx to NULL).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>