python/addrxlat: Track modifications of step.base
Right now, the "base" attribute is calculated. However, since
FullAddress is a mutable object, callers may be tempted to modify
it in-place, e.g.:

  step.base.addr = 0xbadbadbad

This does not work, because Python first creates a new FullAddress
object for the current content of addrxlat_step_t, then modifies
that new object's "addr" attribute and then throws away the modified
object again.

Let's reuse the work I did on Description attributes, so the above
actually works. After the change, Step objects contain a reference
to an actual FullAddress object, which is synchronized with the
addrxlat_step_t structure.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>