python/addrxlat: Fix refcounting in step_Init()
When re-initializing an existing Step object, previous references to
the ctx, sys, and meth objects are leaked.

This is a bit tricky, because ctx should stay in sync with the
ctx object, but there is a call to set_fulladdr() between making
the Python object and loc_scatter (which takes care of updating the
pointers in self->desc). If that call fails, the Step object may end
up in an inconsistent state (where step.ctx refers to another context
than Step.ctx).

One option would have been to provide a mechanism to roll back, but
it is much simpler to swap the two calls, because all other calls are
fail-safe.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>