python/addrxlat: Fix initialization of CustomDescription
The callback data must be overridden to allow modifying the Python
callback later.

However, this fix reveals a conceptual flaw in custom callbacks:
If the additional data is only stored in addrxlat_desc_t, which is
copied into the associated addrxlat_meth_t, then there is no clean
way to modify the pointer for just one call.

So, for now this fix makes all test cases fail until a better
mechanism is invented. *sigh*

Signed-off-by: Petr Tesarik <ptesarik@suse.com>