addrxlat_step_t: Replace the meth pointer with a desc pointer
This commit verifies that a method object is no longer needed.
A temporary object is created in addrxlat_step_* functions to
separate the API change from rewriting the step setup methods,
but it is no longer crucial for the translation.

BTW note (in the test suite) how this change can simplifiy the
use of the library...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>