addrxlat_map_t: Store method index rather than method pointer
The main benefit is that a method index does not have a reference
count (it is a plain non-shareable enum type).

The main issue is that such change of the type ripples up the whole
stack up to Python example scripts. :-( Well, let's do it now while
the API is still considered unstable.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>