python/addrxlat: Store a reference to System in Step
This approach is slightly less confusing for casual users, because
the "sys" attribute now behaves more like a normal attribute.

I mean, changes made to the system through step.sys may or may not
be persistent, depending on whether or not they are reflected in the
underlying C object. With the present implementation, all changes are
preserved.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>