cache: Store cache entry state in its own field
First, these "flags" are in fact state of the cache entry (valid,
pending for probed list, or pending for precious list).

Second, there is no advantage in combining the flags with the pfn
field, because struct cache_entry contains a hole anyway, but the
code is much harder to read when these high bits are reused.

Third, we'll get the full size of kdump_pfn_t for the cache index.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>