cache: Allow zero-size elements
Calling malloc() with a size of 0 is not portable. However, a cache
with zero-size elements may still be useful if the data pointer is
managed by the caller.

Note that NULL cannot be used as cache->data, because then all
elements would have a NULL data pointer, which has a special meaning
to the cache algorithm (it denotes ghost elements). We could use any
non-NULL pointer value here, but let's use the cache address, because
that's unique to every cache, which is potentially helpful for
debugging.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>