fcache: Implement file chunks
A file cache is a crude interface - it has block granularity, and
reading data at arbitrary positions requires a lot of work.

This code adds an abstraction of file chunks. The code could be much
simpler if I always allocated a separate buffer and copied all the
data into that buffer. However, the allocation and copying is not
needed if the whole chunk fits into a single file cache entry, or if
cache data of adjacent entries happens to be contiguous in virtual
memory.

The new code optimizes all these cases, making a copy only if
absolutely necessary.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>