Cleanup: Handle missing ELF pages earlier
The program header list is now always searched to decide if data
can be served from the file cache, or if it must be constructed in
the generic page cache. If and only if no program header can be
found, the page does not exist in the ELF dump. So, let's move the
check here.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>