fcache: Allow zero-sized chunks
The previous commit broke the test suite (which contains empty
diskdump files). This alone would not be a compelling reason to
add more code. However, since the cache chunk interface is
intended as an improvement over read(3) and pread(3) calls, which
succeed if length is zero, callers may want to omit special
treatment of zero-sized buffers.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>