ELF xc_core: Use reverse maps for physical <-> machine translation
Modify the custom methods for converting between physical and
machine addresses in a DomU dump to use the same maps.

I am not adding separate maps for these conversions, because these
operations are not expected to be time-critical, so converting to a
file index and reading the corresponding map entry using file cache
should be sufficient.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>