Cleanup: Remove xen_map_size from shared info
The field is used only during initialization.

Note that the loop termination condition in make_xen_pfn_map_* had
to be adjusted to cope gracefully with a section size that is not a
multiple of the respective element size. Previously, this was
handled by division and multiplication, but it can be also done
simply by comparing against the highest file offset which would
still give a full element.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>