addrxlat/x86_64: Fix ktext fallback to page tables
If the kernel text mapping cannot be determined, os_init should
fall back to using page tables for that range. Unfortunately, this
code is never executed. As a result, there is no translation for
virtual addresses in the kernel text range (without the bugfix,
there is no @ktext: section in the test case output).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>