addrxlat/x86_64: Prefer _stext to set up kernel text mapping
The current code tries some common values for LOAD_ADDRESS. But if
the address of the _stext symbol is known, it should be preferred,
because it is known to be inside the kernel text area.

Note that this method will work even with kASLR, because symbols
found in VMCOREINFO or in /proc/kallsyms map to the actual values,
i.e. after kASLR relocation.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>