addrxlat/x86_64: If _stext is unknown, search the entire kernel text area
With all the recent utility functions, it is now possible to
replace the two most common kernel text locations with a full
search over the entire area. This works with any value of the
PHYSICAL_START config option.

Note that the stub page tables for tests must be extended,
otherwise the search fails, because data cannot be read.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>