addrxlat/x86_64: Do add a default directmap for Linux 4.8+
Linux 4.8 introduced kASLR of the directmap region. In absence of
any other piece of information apart from the kernel version, the
code must assume that randomization was in effect and refuse to
create a directmap (also resulting in no reverse directmap).

It is better to fail with an error than to return wrong information
silently.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>