addrxlat: Fix page table walk
When moving to the next page table entry at a given level, the
search must continue with the first (or last for downward scan)
entry of the lower-level table. However, the search will in fact
continue at an offset corresponding to the search starting address.

The current code happens to work for x86_64 Linux kernel text,
because the search starts at an address that is aligned to a
multiple of the top-level page table entry span. It breaks for an
unaligned address.

The fix includes:

  1. a different method to move the virtual address, and
  2. adjusting the translation step indices.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>