addrxlat/x86_64: Determine the actual size of the kernel text mapping
Walk the page tables to find the end of the kernel text
mapping.

Note that the end of the mapping may be determined incorrectly if
kernel text is followed by module pages which happen to be mapped
to physical address with the same offset as the kernel core.
AFAICS, if this happens, it should make no difference to library
users in practice.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>