Add helper function highest_linear()
The highest linearly mapped address can be determined by combining
lowest_nonlinear() and highest_mapped(). Such calculation should be
useful in more than one place, so let's keep things DRY.

Note that this is an inline function. That's not because of speed,
but it appears that the overhead of a function call would be bigger
than the savings from reusing code, so an inline function is likely
to be both faster _and_ smaller.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>