addrxlat/x86_64: Find the actual extenst of direct physical mapping
Search the page tables to find the actual start and end of the
direct mapping on x86_64. This method works well with kASLR,
because the direct physical mapping is always the first mapping
above LINUX_DIRECTMAP_START_2_6_31.

Note that the test cases had to be vastly extended to include a
complete page table hierarchy...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>