Cleanup: Remove is_mapped()
This function is no longer used for anything. It was previously
used to determine whether VMEMMAP was present at a given virtual
address. This test was necessary to get the size of the direct
mapping. Since direct mapping is now sized by walking the page
tables themselves, the VMEMMAP check was dropped.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>