addrxlat/x86_64: Re-initialize rdirect mapping in linux_directmap()
A temporary mapping may be set up to allow walking the page tables
if only virtual addresses can be used to access the dump file.
This mapping may be inaccurate (e.g. it may be larger, resulting in
virtual addresses that map back to a different physical address).

To avoid any confusion, remove the temporary mapping after initial
bootstrapping.

The direct map entry for kernels between 2.6.27 and 2.6.31 can be
removed, because it only differs from the 2.6.31+ mapping in the
end address, which is not an issue if the mapping is removed later.

The test for non-NULL map in remove_redirect() had to be added,
because at this point there may be no reverse direct map (unlike
the original call site).

I have updated the kvaddr test case to use a smaller mapping than
the full 64T to test if the reverse direct mapping is reduced.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>