addrxlat/x86_64: Always create a maximum-sized temporary rdirect map
The temporary mapping is used only to walk page tables, which
should never contain an out-of-range physical address. If they do,
the data is corrupt, and libaddrxlat cannot be expected to work
properly... Anyway, this assumption simplifies the logic, because
the temporary mapping can always cover all potential physical
addresses. Consequently, there is no need to store the end address
of the direct mapping.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>