Use file cache to read ELF string table
Do not use pread(2) to load the string table, removing last
occurence of direct file access in elfdump.c.

Note that this was also the last user of read_elf_sect(). This
routine is flawed by design, because it returns NULL on allocation
failure or read failure, so there is no way to produce a meaningful
error message.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>