diskdump: Implement file page bitmap
Use the PFN region map to set bitmap bits. I do not copy from the
file bitmap directly, because the target bitmap may not be aligned
to a byte boundary. I specifically do not want to add an alignment
constraint, because kdump_bmp_get_bits() in its present form can be
easily used to implement a check whether a bitmap bit is set (just
use the same value for first and last).

I had to modify find_pfn_rgn() to avoid repeating the search if the
starting PFN happens to be outside of any range. Note that this
requires an additional comparison in pfn_to_pdpos().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>