diskdump: Fix skip_set()
The bitwise negation operator implicitly casts its operand to an
int. If the is an unsigned char, then it is zero-extended first,
but that means that there is always a set bit at position 8 and the
function never skips more than a single byte.

This bug was revealed after looking at dumpattr output for
file.pagemap.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>