addrxlat_op: Do not reset recursion detection
The distinction between internal and external API users is moot.
External users cannot reuse the same context from different threads,
so the only way to request translating the same address through the
same context is by recursion from a callback function. It does not
matter that a separate library call was made. If such recursive
address translation is needed once, it is probably needed every
time.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>