Unify error string handling and move it to a header file
Both libkdumpfile and libaddrxlat implement the same routines for
handling error strings, merely in a different structure. Clean this
up and move the routines to a separate file.

Note that I'm not building a convenience library, because that would
cause all kinds of trouble (sanitizing internal names, fighting with
libtool including the whole archive, etc.).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>