Attribute cloning API
Add an API to clone whole attribute hierarchies. The basic idea is
that a cloned hierarchy uses the base attributes, unless they are
overridden in its own (local) hierarchy.

A cloned hierarchy holds a pointer to the base hierarchy (in its
'fallback' field). Since this increases the reference count, the
cloned hierarchy can freely link to any data from the base
hierarchy, most importantly attribute templates.

The original attribute lookup API is slightly modified. If the key
starts with a dot ('.'), then the fallback hierarchy is not
searched. This is useful to set up overrides in a cloned hierarchy.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>