Allow different address translation in kdump context clones
It is sometimes desirable to create two different kdump contexts for
the same file, but using different address translation. For example,
in a Xen host dump, one context could be used to inspect the Dom0
kernel, and another one to inspect the Xen hypervisor.

The above can be currently achieved by creating two separate
contexts for the same file, but it is suboptimal, because those two
contexts do not share caches and other common data.

Instead, let's add a clone flag to kdump_clone() and give the
library user control over what data should be shared with the
original kdump context.

The test case creates an ELF dump file with two page table
hierarchies and then uses addrxlat.opts attribute to use one or the
other through different contexts, checking that there is no race
over the attribute or the translation definition itself.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>