Split addrxlat.opts to addrxlat.opts.pre and addrxlat.opts.post
This change allows to add custom libaddrxlat options both before and
after the automatically generated options. Since only the last
occurence of an option is taken by libaddrxlat, you can effectively
specify a default (with addrxlat.opts.pre) or an override (with
addrxlat.opts.post), which should increase flexibility in using the
library.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>