Reverse the logic of kdump_clone() bits
The original design was motivated by the clone(2) syscall, but as
more clone flags are likely to be added in the future, it's better
to treat no flags as "clone everything" and add flag bits for parts
that should be separate.

The main reason is that whenever a new bit is added, it would have
to be added to KDUMP_CLONE_ALL, breaking binary compatibility with
programs compiled for an older version.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>